refactore the app class
