+++ app.rb
@@ -1,13 +1,17 @@
 require './book'
 require './student'
 require './teacher'
+require_relative 'create_person'
+require_relative 'person_rentals'
 
 class App
+  include PersonRentals
   attr_accessor :books, :people
 
   def initialize
+    @person = CreatePerson.new
     @books = []
-    @people = []
+    @people = @person.people
   end
 
   def list_books(with_index)
@@ -21,21 +25,8 @@ class App
     @books << Book.new(title, author)
   end
 
-  def add_student(age, name, parent_permession)
-    @people << { type: 'Student', data: Student.new(age, name, parent_permession) }
-  end
-
-  def add_teacher(age, name, specialization)
-    @people << { type: 'Teacher', data: Teacher.new(age, name, specialization) }
-  end
-
   def add_person(type, age, name, third_param)
-    case type
-    when 1
-      add_student(age, name, third_param)
-    when 2
-      add_teacher(age, name, third_param)
-    end
+    @person.create_person(type, age, name, third_param)
   end
 
   def list_people(with_index)
@@ -50,14 +41,7 @@ class App
     @books[book_index].add_rental(@people[person_index][:data], date)
   end
 
-  def find_person(person_id)
-    @people.find { |person| person[:data].id == person_id }
-  end
-
   def list_person_rentals(id)
-    person = find_person(id)
-    person && person[:data].rentals.each do |rental|
-      puts "Date: #{rental.date}, Book: #{rental.book.title} by #{rental.book.author}"
-    end
+    person_rentals(id, @people)
   end
 end
